feat: data load/delete commands 일부 수정
